fix worker

+++ worker/worker.py
@@ -8,7 +8,6 @@ from enum import StrEnum
 
 from aiobotocore.session import get_session
 from langcodes import Language
-from requests import Session
 from sqlalchemy.orm import sessionmaker
 from utils import (
     check_file_in_s3,
@@ -61,7 +60,7 @@ async def check_finish(job: WorkerJob) -> bool:
     )
 
 
-async def process_job(job: WorkerJob, use_gpt4: bool) -> None:
+async def process_job(job: WorkerJob, use_gpt4: bool) -> bool:
     """根据SQS job的类型进行原文翻译/摘要生成/标题翻译的工作."""
     article_path = os.path.join(job.s3_prefix, "article.txt")
     target_lang = Language.get(job.target_lang).display_name()
@@ -94,33 +93,38 @@ async def process_job(job: WorkerJob, use_gpt4: bool) -> None:
         case _:
             raise ValueError(f"Not supported job type: {job.job_type}")
 
-    await put_text_file_to_s3(res_path, text)
+    if text != "":
+        await put_text_file_to_s3(res_path, text)
+
+    return text != ""
 
 
-def put_user_articles(
-    media_id: int, article_id: int, lang: str, db_session: sessionmaker[Session]
-) -> None:
+def put_user_articles(media_id: int, article_id: int, lang: str) -> None:
     """完成文章的摘要和翻译后将其推送给已订阅的用户."""
-    session = db_session()
+    engine = get_db_engine()
+    db_sess = sessionmaker(bind=engine)()
     try:
         user_medias = (
-            session.query(UserMedia)
+            db_sess.query(UserMedia)
             .filter(UserMedia.media_id == media_id and UserMedia.lang == lang)
             .all()
         )
+        user_articles = []
         for user_media in user_medias:
-            session.add(UserArticle(user_id=user_media.user_id, article_id=article_id))
+            user_articles.append(
+                UserArticle(user_id=user_media.user_id, article_id=article_id)
+            )
 
-        session.commit()
+        db_sess.add_all(user_articles)
+        db_sess.commit()
     finally:
-        session.close()
+        db_sess.close()
+        engine.dispose()
 
 
 async def main(args: argparse.Namespace) -> None:
-    session = get_session()
-    engine = get_db_engine()
-    sessionmaker(bind=engine)
-    async with session.create_client("sqs", region_name="us-west-2") as sqs:
+    aio_sess = get_session()
+    async with aio_sess.create_client("sqs", region_name="us-west-2") as sqs:
         while True:
             try:
                 response = await sqs.receive_message(
@@ -133,18 +137,20 @@ async def main(args: argparse.Namespace) -> None:
                         logging.info(f"process job: {job}")
                         if job.job_type == JobType.GET_IMAGE:
                             continue
-                        await process_job(job, args.use_gpt4)
-                        await sqs.delete_message(
-                            QueueUrl=QUEUE_WORKER_URL,
-                            ReceiptHandle=message["ReceiptHandle"],
-                        )
-                        if await check_finish(job):
-                            logging.info(
-                                f"All jobs of the article {job.article_id} is finished!"
+                        if await process_job(job, args.use_gpt4):
+                            await sqs.delete_message(
+                                QueueUrl=QUEUE_WORKER_URL,
+                                ReceiptHandle=message["ReceiptHandle"],
                             )
-                            put_user_articles(
-                                job.media_id, job.article_id, job.target_lang
-                            )
+                            if await check_finish(job):
+                                logging.info(
+                                    f"All jobs of the article {job.article_id} is finished!"
+                                )
+                                put_user_articles(
+                                    job.media_id,
+                                    job.article_id,
+                                    job.target_lang,
+                                )
                 else:
                     logging.info("No messages in queue.")
             except Exception as e:
